feat: sync url to scoreboard params
